backend/requirements.txt
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cmake==3.26.1
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
data-science-types==0.2.23
debugpy==1.6.6
decorator==5.1.1
Django==4.1.7
django-cors-headers==3.14.0
django-stubs==1.16.0
django-stubs-ext==0.8.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djangorestframework-stubs==1.10.0
executing==1.2.0
filelock==3.10.7
fonttools==4.39.3
huggingface-hub==0.13.3
idna==3.4
ipykernel==6.22.0
ipython==8.12.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.1.0
jupyter_core==5.3.0
kiwisolver==1.4.4
lit==16.0.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mpmath==1.3.0
mypy==1.1.1
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==3.0
nltk==3.8.1
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
packaging==23.0
pandas==1.5.3
pandas-stubs==1.5.3.230321
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.2.0
prompt-toolkit==3.0.38
psutil==5.9.4
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
PyJWT==2.6.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
pyzmq==25.0.2
regex==2023.3.23
requests==2.28.2
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
sentence-transformers==2.2.2
sentencepiece==0.1.97
six==1.16.0
sqlparse==0.4.3
stack-data==0.6.2
sympy==1.11.1
threadpoolctl==3.1.0
tokenizers==0.13.2
tomli==2.0.1
torch==2.0.0
torchvision==0.15.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.27.4
triton==2.0.0
types-docutils==0.19.1.7
types-Markdown==3.4.2.7
types-openpyxl==3.1.0.3
types-Pillow==9.4.0.19
types-psycopg2==2.9.21.9
types-Pygments==2.15.0.0
types-pytz==2023.3.0.0
types-PyYAML==6.0.12.9
types-requests==2.28.11.17
types-setuptools==67.6.0.7
types-six==1.16.21.8
types-urllib3==1.26.25.10
typing_extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6
